Second Solution/requirements.txt
alembic==1.8.1
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
async-generator==1.10
attrs==22.1.0
Babel==2.10.3
backcall==0.2.0
backoff==1.11.1
beautifulsoup4==4.11.1
bleach==5.0.1
blis==0.7.8
brotlipy==0.7.0
catalogue==2.0.8
certifi==2022.6.15
certipy==0.1.3
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1631636250774/work
chardet==3.0.4
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1626371162869/work
click==8.1.3
cloudpickle==2.1.0
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1602866480661/work
conda==4.10.3
conda-package-handling @ file:///home/conda/feedstock_root/build_artifacts/conda-package-handling_1618231390031/work
cryptography @ file:///home/conda/feedstock_root/build_artifacts/cryptography_1633983255347/work
cycler==0.11.0
cymem==2.0.6
databricks-cli==0.17.2
DAWG-Python==0.7.2
debtcollector==2.5.0
debugpy==1.6.3
decorator==5.1.1
deep-translator==1.8.3
defusedxml==0.7.1
docker==6.0.0
docopt==0.6.2
encoder @ file:///opt/extensions/encoder
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.10.0
fastjsonschema==2.16.1
Flask==2.2.2
fonttools==4.37.1
fuzzywuzzy==0.18.0
gensim==4.2.0
gitdb==4.0.9
GitPython==3.1.27
google-trans-new==1.1.9
greenlet==1.1.3
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2022.8.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib==1.0.4
importlib-metadata==4.12.0
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.2
iso8601==1.0.2
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
json5==0.9.10
jsonschema==4.14.0
jupyter-client==7.3.4
jupyter-core==4.11.1
jupyter-resource-usage==0.6.2
jupyter-server==1.18.1
jupyter-telemetry==0.1.0
# Editable install with no version control (jupyterhub==2.2.0.dev0)
-e /opt/jupyterhub
jupyterlab==3.4.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
jupyterlab_server==2.15.1
keystoneauth1==5.0.0
kiwisolver==1.4.4
langcodes==3.3.0
libcst==0.4.7
lxml==4.9.1
Mako==1.2.1
mamba @ file:///home/conda/feedstock_root/build_artifacts/mamba_1632770295204/work
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mistune==2.0.4
mlflow @ file:///opt/mlflow-1.26.1.dev0-py3-none-any.whl
msgpack==1.0.4
murmurhash==1.0.8
mypy-extensions==0.4.3
nbclassic==0.4.3
nbclient==0.6.7
nbconvert==7.0.0
nbformat==5.4.0
nbgitpuller==1.1.0
nest-asyncio==1.5.5
netaddr==0.8.0
netifaces==0.11.0
nltk==3.7
notebook==6.4.12
notebook-shim==0.1.0
nteract-on-jupyter==2.1.3
numpy==1.23.2
oauthlib==3.2.0
openpyxl==3.0.10
os-service-types==1.7.0
oslo.config==9.0.0
oslo.i18n==5.1.0
oslo.serialization==5.0.0
oslo.utils==6.0.1
outcome==1.2.0
packaging==21.3
pamela==1.0.0
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pathy==0.6.2
pbr==5.10.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
plotly==5.10.0
positional==1.2.1
preshed==3.0.7
prometheus-client==0.14.1
prometheus-flask-exporter==0.20.3
prompt-toolkit==3.0.30
protobuf==4.21.5
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pycosat @ file:///home/conda/feedstock_root/build_artifacts/pycosat_1610094799048/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1593275161868/work
pydantic==1.9.2
pyenchant==3.2.2
Pygments==2.13.0
PyJWT==2.4.0
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pymystem3==0.2.0
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1633192417276/work
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1610291451001/work
python-dateutil==2.8.2
python-json-logger==2.0.4
python-keystoneclient==3.10.0
python-Levenshtein==0.12.2
pytz==2022.2.1
PyVirtualDisplay==3.0
PyYAML==6.0
pyzmq==23.2.1
querystring-parser==1.2.4
regex==2022.8.17
requests==2.28.1
rfc3986==1.5.0
ru-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/ru_core_news_sm-3.4.0/ru_core_news_sm-3.4.0-py3-none-any.whl
ruamel-yaml-conda @ file:///home/conda/feedstock_root/build_artifacts/ruamel_yaml_1611943432947/work
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-learn==1.1.2
scipy==1.9.0
seaborn==0.11.2
selenium==4.4.3
selenium-stealth==1.0.6
Send2Trash==1.8.0
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sklearn==0.0
smart-open==5.2.1
smmap==5.0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spacy==3.4.1
spacy-legacy==3.0.10
spacy-loggers==1.0.3
SQLAlchemy==1.4.40
sqlparse==0.4.2
srsly==2.4.4
stack-data==0.4.0
stevedore==4.0.0
tabulate==0.8.10
tenacity==8.0.1
terminado==0.15.0
thinc==8.1.0
threadpoolctl==3.1.0
tinycss2==1.1.1
toml==0.10.2
tornado==6.2
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1632160078689/work
traitlets==5.3.0
transliterate==1.10.2
trio==0.21.0
trio-websocket==0.9.2
typer==0.4.2
typing-inspect==0.8.0
typing_extensions==4.3.0
unimport==0.11.1
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1632350318291/work
wasabi==0.10.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.0
Werkzeug==2.2.2
widgetsnbextension==3.6.1
word2vec==0.11.1
wrapt==1.14.1
wsproto==1.2.0
zipp==3.8.1
python -m gensim.downloader --download word2vec-ruscorpora-300